requirements.txt
streamlit
groq
python-dotenv
langchain-community
langchain-text-splitters
sentence-transformers
chromadb